+++ src/app/admin/invite/page.tsx
@@ -2,7 +2,7 @@ import { Title, Container } from '@mantine/core'
 import { pageStyles, mainStyles, footerStyles } from '@/styles/common'
 import { OrganizationSelect } from '@/components/clerkcustom/organization-select'
 import { getAllOrganizations } from '@/server/actions/organization-actions'
-import InviteForm from '@/components/invite-form'
+import InviteForm from '@/components/clerkcustom/invite-form'
 
 export default async function InvitePage() {
     const organizations = await getAllOrganizations()
@@ -12,8 +12,7 @@ export default async function InvitePage() {
             <main className={mainStyles}>
                 <Container size="sm">
                     <Title mb="lg">Invite New Users</Title>
-                    <OrganizationSelect organizations={organizations} />
-                    <InviteForm />
+                    <InviteForm organizations={organizations} />
                 </Container>
             </main>
         </div>
